cleanup and remove compiled files
