+++ src/abci/handlers/order.ts
@@ -4,7 +4,7 @@
   orderHandlers.ts @ {master}
   =========================
 
-  @date_inital 23 October 2018
+  @date_initial 23 October 2018
   @date_modified 29 October 2018
   @author Henry Harder
 
@@ -13,11 +13,11 @@
 
 import * as Paradigm from "paradigm-connect";
 
-import { messages as msg } from "../util/messages";
-import { Logger } from "../util/Logger";
-import { Vote } from "../util/Vote";
-import { Hasher } from "../crypto/Hasher";
-import { OrderTracker } from "../async/OrderTracker";
+import { messages as msg } from "../../util/messages";
+import { Logger } from "../../util/Logger";
+import { Vote } from "../../util/Vote";
+import { Hasher } from "../../crypto/Hasher";
+import { OrderTracker } from "../../async/OrderTracker";
 
 // Order constructor from paradigm-connect
 let Order = new Paradigm().Order;
