+++ src/crypto/PayloadCipher.ts
@@ -4,7 +4,7 @@
   PayloadCipher.ts @ {master}
   =========================
 
-  @date_inital 21 September 2018
+  @date_initial 21 September 2018
   @date_modified 19 October 2018
   @author Henry Harder
