+++ src/net/WebSocketMessage.ts
@@ -4,7 +4,7 @@
   WebSocketMessage.ts @ {master}
   =========================
   
-  @date_inital 27 August 2018
+  @date_initial 27 August 2018
   @date_modified 19 October 2018
   @author Henry Harder
