+++ src/state/deliverState.ts
@@ -4,7 +4,7 @@
   deliverState.ts @ {master}
   =========================
   
-  @date_inital 25 September 2018
+  @date_initial 25 September 2018
   @date_modified 24 October 2018
   @author Henry Harder
