+++ src/util/Logger.ts
@@ -4,7 +4,7 @@
   Logger.ts @ {master}
   =========================
 
-  @date_inital 25 September 2018
+  @date_initial 25 September 2018
   @date_modified 19 October 2018
   @author Henry Harder
